Add dates to /press/ page, and sort by date, recent-first
Looking at other "press coverage" pages, they tend to list
the date that the press described the event. This is helpful
for visitors who get some context about the link, but moreover
gives the press page the feeling of recency, suggesting that
press interest (and the event) is still active.