Move "How the event has spread" to /press/
One of the key things that the press often does not know about
our workshop is that it is the original event in the Python
community, from which many others were inspired. This commit
moves that information to the press page.

It also adds dates to that section to show that these collaborations
started many months ago and are continuing apace.

The "run your own" retains a link to the press page so that
prospective collaborators see that information as well.